fix download on inter b2b courses for coach
